Added tests for Forms\Form.
