+++ tests/integration/Forms/Form/AddCest.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 namespace Phalcon\Test\Integration\Forms\Form;
 
 use IntegrationTester;
+use Phalcon\Forms\Element\Text;
+use Phalcon\Forms\Form;
 
 /**
  * Class AddCest
@@ -31,5 +33,25 @@ class AddCest
     {
         $I->wantToTest('Forms\Form - add()');
         $I->skipTest('Need implementation');
+    }
+
+    /**
+     * @issue https://github.com/phalcon/cphalcon/issues/706
+     */
+    public function testIssue706(IntegrationTester $I)
+    {
+        $form = new Form();
+        $form->add(new Text("name"));
+
+        $form->add(new Text("before"), "name", true);
+        $form->add(new Text("after"), "name");
+
+        $expected = ["before", "name", "after"];
+        $actual   = [];
+
+        foreach ($form as $element) {
+            $actual[] = $element->getName();
+        }
+        $I->assertEquals($expected, $actual);
     }
 }
