+++ tests/integration/Forms/FormCest.php
@@ -43,27 +43,6 @@ class FormCest
         Tag::setDocType(Tag::XHTML5);
     }
 
-    public function testCount(IntegrationTester $I)
-    {
-        $form = new Form();
-
-        $expected = 0;
-        $actual   = count($form);
-        $I->assertEquals($expected, $actual);
-
-        $form->add(
-            new Text("name")
-        );
-
-        $form->add(
-            new Text("telephone")
-        );
-
-        $expected = 2;
-        $actual   = count($form);
-        $I->assertEquals($expected, $actual);
-    }
-
     public function testIterator(IntegrationTester $I)
     {
         $form = new Form();
@@ -133,104 +112,6 @@ class FormCest
         $expected = "<label for=\"telephone\" class=\"form-control\">The Telephone</label>";
         $actual   = $form->label("telephone", ["class" => "form-control"]);
         $I->assertEquals($expected, $actual);
-    }
-
-    /**
-     * Tests Form::hasMessagesFor
-     *
-     * @author Sid Roberts <Sid@SidRoberts.co.uk>
-     * @since  2016-04-03
-     */
-    public function testFormHasMessagesFor(IntegrationTester $I)
-    {
-        // First element
-        $telephone = new Text('telephone');
-        $telephone->addValidators(
-            [
-                new Regex(
-                    [
-                        'pattern' => '/\+44 [0-9]+ [0-9]+/',
-                        'message' => 'The telephone has an invalid format',
-                    ]
-                ),
-            ]
-        );
-
-        // Second element
-        $address = new Text('address');
-        $form    = new Form();
-
-        $form->add($telephone);
-        $form->add($address);
-
-        $actual = $form->isValid(['telephone' => '12345', 'address' => 'hello']);
-        $I->assertFalse($actual);
-
-        $expected = new Messages(
-            [
-                new Message(
-                    'The telephone has an invalid format',
-                    'telephone',
-                    'Regex',
-                    0
-                ),
-            ]
-        );
-        $actual   = $form->getMessagesFor('telephone');
-        $I->assertEquals($expected, $actual);
-
-        $expected = new Messages();
-        $actual   = $form->getMessagesFor('address');
-        $I->assertEquals($expected, $actual);
-
-        $actual = $form->hasMessagesFor('telephone');
-        $I->assertTrue($actual);
-
-        $actual = $form->hasMessagesFor('address');
-        $I->assertFalse($actual);
-    }
-
-    /**
-     * Tests Form::render
-     *
-     * @issue  https://github.com/phalcon/cphalcon/issues/10398
-     * @author Phalcon Team <team@phalconphp.com>
-     * @since  2016-07-17
-     */
-    public function testCreatingElementsWithNameSimilarToTheFormMethods(IntegrationTester $I)
-    {
-        $names = [
-            'validation',
-            'action',
-            'useroption',
-            'useroptions',
-            'entity',
-            'elements',
-            'messages',
-            'messagesfor',
-            'label',
-            'value',
-            'di',
-            'eventsmanager',
-        ];
-
-        foreach ($names as $name) {
-            $form    = new Form;
-            $element = new Text($name);
-
-            $expected = $name;
-            $actual   = $element->getName();
-            $I->assertEquals($expected, $actual);
-
-            $form->add($element);
-
-            $expected = sprintf('<input type="text" id="%s" name="%s" />', $name, $name);
-            $actual   = $form->render($name);
-            $I->assertEquals($expected, $actual);
-
-            $actual = $form->getValue($name);
-            $I->assertNull($actual);
-        }
     }
 
     public function testFormValidator(IntegrationTester $I)
@@ -352,21 +233,6 @@ class FormCest
         $I->assertTrue($actual);
     }
 
-    public function testFormIndirectElementRender(IntegrationTester $I)
-    {
-        $form = new Form();
-
-        $form->add(new Text("name"));
-
-        $expected = '<input type="text" id="name" name="name" />';
-        $actual   = $form->render("name");
-        $I->assertEquals($expected, $actual);
-
-        $expected = '<input type="text" id="name" name="name" class="big-input" />';
-        $actual   = $form->render("name", ["class" => "big-input"]);
-        $I->assertEquals($expected, $actual);
-    }
-
     /**
      * @issue https://github.com/phalcon/cphalcon/issues/1190
      */
@@ -380,26 +246,6 @@ class FormCest
 
         $actual   = $form->render("title");
         $expected = '<input type="text" id="title" name="title" value="Hello &quot;world!&quot;" />';
-        $I->assertEquals($expected, $actual);
-    }
-
-    /**
-     * @issue https://github.com/phalcon/cphalcon/issues/706
-     */
-    public function testIssue706(IntegrationTester $I)
-    {
-        $form = new Form();
-        $form->add(new Text("name"));
-
-        $form->add(new Text("before"), "name", true);
-        $form->add(new Text("after"), "name");
-
-        $expected = ["before", "name", "after"];
-        $actual   = [];
-
-        foreach ($form as $element) {
-            $actual[] = $element->getName();
-        }
         $I->assertEquals($expected, $actual);
     }
 
@@ -463,205 +309,6 @@ class FormCest
 
         $expected = new Messages();
         $actual   = $form->getMessagesFor('notelement');
-        $I->assertEquals($expected, $actual);
-    }
-
-    /**
-     * Tests Form::setValidation()
-     *
-     * @author Mohamad Rostami <rostami@outlook.com>
-     * @issue  https://github.com/phalcon/cphalcon/issues/12465
-     */
-    public function testCustomValidation(IntegrationTester $I)
-    {
-        // First element
-        $telephone        = new Text('telephone');
-        $customValidation = new Validation();
-        $customValidation->add(
-            'telephone',
-            new Regex(
-                [
-                    'pattern' => '/\+44 [0-9]+ [0-9]+/',
-                    'message' => 'The telephone has an invalid format',
-                ]
-            )
-        );
-        $form    = new Form();
-        $address = new Text('address');
-        $form->add($telephone);
-        $form->add($address);
-        $form->setValidation($customValidation);
-
-        $actual = $form->isValid(['telephone' => '12345', 'address' => 'hello']);
-        $I->assertFalse($actual);
-
-        $actual = $form->get('telephone')->hasMessages();
-        $I->assertTrue($actual);
-
-        $actual = $form->get('address')->hasMessages();
-        $I->assertFalse($actual);
-
-        $expected = new Messages(
-            [
-                new Message(
-                    'The telephone has an invalid format',
-                    'telephone',
-                    'Regex',
-                    0
-                ),
-            ]
-        );
-        $actual   = $form->get('telephone')->getMessages();
-        $I->assertEquals($expected, $actual);
-
-        $expected = $form->getMessages();
-        $actual   = $form->get('telephone')->getMessages();
-        $I->assertEquals($expected, $actual);
-
-        $expected = new Messages();
-        $actual   = $form->get('address')->getMessages();
-        $I->assertEquals($expected, $actual);
-    }
-
-    /**
-     * Tests Form::isValid()
-     *
-     * @author Mohamad Rostami <rostami@outlook.com>
-     * @issue  https://github.com/phalcon/cphalcon/issues/11500
-     */
-    public function testMergeValidators(IntegrationTester $I)
-    {
-        // First element
-        $telephone = new Text('telephone');
-        $telephone->addValidators(
-            [
-                new PresenceOf(
-                    [
-                        'message' => 'The telephone is required',
-                    ]
-                ),
-            ]
-        );
-        $customValidation = new Validation();
-        $customValidation->add(
-            'telephone',
-            new Regex(
-                [
-                    'pattern' => '/\+44 [0-9]+ [0-9]+/',
-                    'message' => 'The telephone has an invalid format',
-                ]
-            )
-        );
-        $form    = new Form();
-        $address = new Text('address');
-        $form->add($telephone);
-        $form->add($address);
-        $form->setValidation($customValidation);
-
-        $actual = $form->isValid(['address' => 'hello']);
-        $I->assertFalse($actual);
-
-        $actual = $form->get('telephone')->hasMessages();
-        $I->assertTrue($actual);
-
-        $actual = $form->get('address')->hasMessages();
-        $I->assertFalse($actual);
-
-        $expected = new Messages(
-            [
-                new Message(
-                    'The telephone has an invalid format',
-                    'telephone',
-                    'Regex',
-                    0
-                ),
-                new Message(
-                    'The telephone is required',
-                    'telephone',
-                    'PresenceOf',
-                    0
-                ),
-            ]
-        );
-        $actual   = $form->get('telephone')->getMessages();
-        $I->assertEquals($expected, $actual);
-
-        $expected = $form->getMessages();
-        $actual   = $form->get('telephone')->getMessages();
-        $I->assertEquals($expected, $actual);
-
-        $expected = new Messages();
-        $actual   = $form->get('address')->getMessages();
-        $I->assertEquals($expected, $actual);
-    }
-
-    /**
-     * Tests Form::getMessages(true)
-     *
-     * @author Mohamad Rostami <rostami@outlook.com>
-     * @issue  https://github.com/phalcon/cphalcon/issues/13294
-     *
-     * This should be removed in next major version
-     * We should not return multiple type of result in a single method!
-     * (form->getMessages(true) vs form->getMessages())
-     */
-    public function testGetElementMessagesFromForm(IntegrationTester $I)
-    {
-        // First element
-        $telephone = new Text('telephone');
-        $telephone->addValidators(
-            [
-                new PresenceOf(
-                    [
-                        'message' => 'The telephone is required',
-                    ]
-                ),
-            ]
-        );
-        $customValidation = new Validation();
-        $customValidation->add(
-            'telephone',
-            new Regex(
-                [
-                    'pattern' => '/\+44 [0-9]+ [0-9]+/',
-                    'message' => 'The telephone has an invalid format',
-                ]
-            )
-        );
-        $form    = new Form();
-        $address = new Text('address');
-        $form->add($telephone);
-        $form->add($address);
-        $form->setValidation($customValidation);
-
-        $actual = $form->isValid(['address' => 'hello']);
-        $I->assertFalse($actual);
-
-        $expected = [
-            'telephone' => [
-                new Messages(
-                    [
-                        new Message(
-                            'The telephone has an invalid format',
-                            'telephone',
-                            'Regex',
-                            0
-                        ),
-                    ]
-                ),
-                new Messages(
-                    [
-                        new Message(
-                            'The telephone is required',
-                            'telephone',
-                            'PresenceOf',
-                            0
-                        ),
-                    ]
-                ),
-            ],
-        ];
-        $actual   = $form->getMessages(true);
         $I->assertEquals($expected, $actual);
     }
 }
